# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AGC Cheswerta CMS API",
    "version": "1.0.0",
    "description": "Church management system API \u2014 members, givings, attendance, fellowship scheduling, council roster, SMS broadcast, and admin tools.\n\n**Getting started:** call `POST /api/auth/login`, copy the `token` from the response, then click **Authorize** above and paste it in as `<token>` (no need to type \"Bearer \" \u2014 Swagger adds that automatically). Every other request will then include it.\n\n**Not included:** a few blueprints registered in the app (announcements, modules, feature flags, generic settings, audit log, account locks, login attempts, file uploads, generic db query, stats/PDF reports, SMS logs, church info) aren't documented here yet \u2014 ask to have them added and share those route files.\n\n**Server selector:** the dropdown at the top switches between Local and Production \u2014 edit the Production URL in the spec to match your real domain."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev server (default Flask port)"
    },
    {
      "url": "https://REPLACE-WITH-YOUR-PRODUCTION-DOMAIN",
      "description": "Production \u2014 replace with your real API domain"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the token from POST /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "You do not have permission to perform this action."
          }
        }
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Attendance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "event_name": {
            "type": "string",
            "example": "Sunday Service"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "men": {
            "type": "integer"
          },
          "women": {
            "type": "integer"
          },
          "youths": {
            "type": "integer"
          },
          "children": {
            "type": "integer"
          },
          "visitors": {
            "type": "integer"
          },
          "total_present": {
            "type": "integer"
          },
          "total_absent": {
            "type": "integer"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "recorded_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AttendanceInput": {
        "type": "object",
        "required": [
          "service_type"
        ],
        "properties": {
          "service_type": {
            "type": "string",
            "example": "Sunday Service",
            "description": "Also accepted as 'event_name'"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "men": {
            "type": "integer",
            "default": 0
          },
          "women": {
            "type": "integer",
            "default": 0
          },
          "youths": {
            "type": "integer",
            "default": 0
          },
          "children": {
            "type": "integer",
            "default": 0
          },
          "visitors": {
            "type": "integer",
            "default": 0
          },
          "total_absent": {
            "type": "integer",
            "default": 0
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/attendance": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "List attendance records",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Record attendance",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attendance"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/attendance/{record_id}": {
      "put": {
        "tags": [
          "Attendance"
        ],
        "summary": "Update an attendance record",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attendance"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Attendance"
        ],
        "summary": "Delete an attendance record (admin/super_admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}